+++ test/analyzers/lint_analyzer/rules/rule_utils_test.dart
@@ -82,25 +82,6 @@ void main() {
       expect(doc2.removeLast(), equals(RuleType.angular.value));
     });
 
-    test('readSeverity returns a Severity from Map based config', () {
-      expect(
-        [
-          {'severity': 'ERROR'},
-          {'severity': 'wArnInG'},
-          {'severity': 'performance'},
-          {'severity': ''},
-          {'': null},
-        ].map((data) => readSeverity(data, Severity.style)),
-        equals([
-          Severity.error,
-          Severity.warning,
-          Severity.performance,
-          Severity.none,
-          Severity.style,
-        ]),
-      );
-    });
-
     group('readExcludes', () {
       test('returns a list of excludes', () {
         const excludes = [